praktikum/week1-intro-cia/requirements.txt
pycryptodome==3.23.0
python-dotenv==1.1.1
